Update the Docker image to take into account the change of the implementation of the JCR from Jackrabbit 2 to Jackrabbit Oak.
The Docker image will embed a JCR migration script that will be used
to migrate existing JCR of previous versions of the container.

By the way, the Silverpeas global configuration file, config.properties,
isn't anymore overridden by a the running of a new version of a
container; it will be check whether there is an already existing
config.properties file.